requirements.txt
# torch>=2.0.0
# Note: SAM3 requires transformers from GitHub main branch (not released yet)
# Install with: pip install git+https://github.com/huggingface/transformers
transformers>=4.40.0
pillow>=10.0.0
numpy==1.26
matplotlib>=3.7.0
opencv-python>=4.8.0
einops>=0.8.0
decord>=0.6.0
timm>=1.0.17
tqdm
ftfy==6.1.1
regex
iopath>=0.1.10
typing_extensions
huggingface_hub
pycocotools>=2.0.0
psutil>=5.0.0
streamlit>=1.28.0
scikit-image>=0.20.0
scikit-learn>=1.0.0
accelerate>=0.20.0

# RunPod SDK for remote GPU support
runpod>=1.6.0
python-dotenv>=1.0.0
requests>=2.31.0
